Moved the marker to its own component.
